requirements.txt
asyncpg>=0.29.0
pgvector>=0.3.6
httpx>=0.27.0
mcp>=1.0.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.34.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
